+++ lib/pbio/sys/command.c
@@ -4,6 +4,7 @@
 #include <assert.h>
 #include <stdint.h>
 
+#include <pbdrv/reset.h>
 #include <pbio/protocol.h>
 
 #include "program_load.h"
@@ -34,6 +35,9 @@ pbio_pybricks_error_t pbsys_command(const uint8_t *data, uint32_t size) {
         case PBIO_PYBRICKS_COMMAND_WRITE_USER_RAM:
             return pbio_pybricks_error_from_pbio_error(pbsys_program_load_set_program_data(
                 pbio_get_uint32_le(&data[1]), &data[5], size - 5));
+        case PBIO_PYBRICKS_COMMAND_REBOOT_TO_UPDATE_MODE:
+            pbdrv_reset(PBDRV_RESET_ACTION_RESET_IN_UPDATE_MODE);
+            return PBIO_PYBRICKS_ERROR_OK;
         default:
             return PBIO_PYBRICKS_ERROR_INVALID_COMMAND;
     }